changing filter

+++ app/checkpoint/antispam.go
@@ -142,10 +142,7 @@ func UrlFilter(context telebot.Context) error {
 			if result.Error != nil {
 				return nil
 			}
-			if link.Type == "blacklist" {
-				return context.Delete()
-			}
-			if GetSpamChance(*context.Sender()) > 50 && result.RowsAffected == 0 {
+			if GetSpamChance(*context.Sender()) > 50 && result.RowsAffected == 0 && link.Type == "blacklist" {
 				return context.Delete()
 			}
 		}
